remove `Mu` from failpoint
It isn't safe to lock each failpoint. The server might panic before
sending response back to client. For example, when a user set a
terms for failpoint A, but the server might be panicking due to
failpoint B. So it isn't safe to lock each single failpoint once
a time.

Instead, we should lock the global mutex. When each failpoint is
triggered, it just needs to acquire the read lock. But when a client
set terms for failpoints, they need to acquire the write lock, and
release the lock after the server(runtime) sends response back to
the client.

Signed-off-by: Benjamin Wang <wachao@vmware.com>